Restructure the validator and util methods
